requirement.txt
python-dotenv==0.19.4
langchain==1.0.0
pinecone-client==2.0.2
streamlit==1.8.0
langchain-community==1.0.0


